requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.5
Django==3.0.7
django-extensions==2.2.9
djangorestframework==3.11.0
future==0.18.2
importlib-metadata==1.6.1
kombu==4.6.10
mysqlclient==1.4.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
uuid==1.30
vine==1.3.0
zipp==3.1.0
